fix sticky navigation bar
The height line was initially added for pages which had
content that was too short to fill the page. The effect was
to cause the footer to appear at the bottom of the page
rather than appearing offset. However, the inclusion of the
height line causes a problem where the navigation bar is
no longer sticky when scrolling past the first row. I have
not found a combination of CSS attributes which handles
both cases, however, there are no longer any short pages
which are affected by this issue right now.